use int64 for gray code

+++ epi/gray_code/solution.go
@@ -1,6 +1,6 @@
 package gray_code
 
-func GrayCode(numBits int) []int {
+func GrayCode(numBits int) []int64 {
 	// TODO - Add your code here
 	return nil
 }
